Moving flask specific code out of common.
