+++ backend/corpora/lambdas/api/v1/dataset.py
@@ -3,7 +3,7 @@ from flask import make_response, jsonify, g
 from ....common.corpora_orm import CollectionVisibility, DatasetArtifactFileType
 from ....common.entities import Dataset, Collection
 from ....common.entities.geneset import GenesetDatasetLink
-from ....common.utils.db_session import dbconnect
+from ....api_server.db import dbconnect
 from ....common.utils.exceptions import (
     NotFoundHTTPException,
     ServerErrorHTTPException,
